Add Shodan in Unclassified category.
Shodan: Shodan is often referred to as a search engine for internet-connected devices. It allows you to explore and search for various devices, including webcams, routers, servers, and more, that are connected to the internet. Shodan provides detailed information about these devices, making it a unique resource for cybersecurity researchers, network administrators, and curious individuals.