Update README.md
1. Unique Content and Variety of Cameras:
AlleCam offers unique content that covers various locations around the world. The webcams show scenic views, cultural landmarks, natural landscapes and city panoramas, making it attractive to a wide audience of users who want to see the world in real time.

2. Regular Updates and Support:
Our site is constantly updated with new cameras and relevant information.

3. User-friendly and Intuitive Interface:
AlleCam offers a user-friendly and intuitive interface that allows users to easily find the cameras they are interested in and enjoy the broadcasts. Our site is adapted for various devices, making it accessible for both desktop computers and mobile devices.

4. Multilingual Support:
AlleCam is available in several languages, allowing users from all over the world to use it comfortably. This expands our audience and makes the site attractive to international users.

Adding the AlleCam website to the section of websites with webcams will enrich the existing list of resources, giving users the opportunity to enjoy high-quality and diverse broadcasts from all over the world.